Create DateNavigator component
